Refactor LegacyInterceptorProviderTests to unit tests and move them out of RequestChainTests
