+++ Tests/ApolloTests/RequestChainTests.swift
@@ -12,66 +12,7 @@ import ApolloTestSupport
 import StarWarsAPI
 
 class RequestChainTests: XCTestCase {
-  
-  lazy var legacyClient: ApolloClient = {
-    let url = TestURL.starWarsServer.url
-    let store = ApolloStore()
-    let provider = LegacyInterceptorProvider(store: store)
-    let transport = RequestChainNetworkTransport(interceptorProvider: provider,
-                                                 endpointURL: url)
-    
-    return ApolloClient(networkTransport: transport, store: store)
-  }()
-  
-  func testLoading() {
-    let expectation = self.expectation(description: "loaded With legacy client")
-    legacyClient.fetch(query: HeroNameQuery()) { result in
-      switch result {
-      case .success(let graphQLResult):
-        XCTAssertEqual(graphQLResult.source, .server)
-        XCTAssertEqual(graphQLResult.data?.hero?.name, "R2-D2")
-      case .failure(let error):
-        XCTFail("Unexpected error: \(error)")
-        
-      }
-      expectation.fulfill()
-    }
-    
-    self.wait(for: [expectation], timeout: 10)
-  }
-  
-  func testInitialLoadFromNetworkAndSecondaryLoadFromCache() {
-    let initialLoadExpectation = self.expectation(description: "loaded With legacy client")
-    legacyClient.fetch(query: HeroNameQuery()) { result in
-      switch result {
-      case .success(let graphQLResult):
-        XCTAssertEqual(graphQLResult.source, .server)
-        XCTAssertEqual(graphQLResult.data?.hero?.name, "R2-D2")
-      case .failure(let error):
-        XCTFail("Unexpected error: \(error)")
-        
-      }
-      initialLoadExpectation.fulfill()
-    }
-    
-    self.wait(for: [initialLoadExpectation], timeout: 10)
-    
-    let secondLoadExpectation = self.expectation(description: "loaded With legacy client")
-    legacyClient.fetch(query: HeroNameQuery()) { result in
-      switch result {
-      case .success(let graphQLResult):
-        XCTAssertEqual(graphQLResult.source, .cache)
-        XCTAssertEqual(graphQLResult.data?.hero?.name, "R2-D2")
-      case .failure(let error):
-        XCTFail("Unexpected error: \(error)")
-        
-      }
-      secondLoadExpectation.fulfill()
-    }
-    
-    self.wait(for: [secondLoadExpectation], timeout: 10)
-  }
-  
+
   func testEmptyInterceptorArrayReturnsCorrectError() {
     class TestProvider: InterceptorProvider {
       func interceptors<Operation: GraphQLOperation>(for operation: Operation) -> [ApolloInterceptor] {
